Adjust the folder structure
