+++ src/components/dictionary/index.tsx
