--- src/features/dictionary/dictionary.d.ts
@@ -1,3 +1,0 @@
-export interface IHeteronyms {
-  heteronyms: [];
-}
