--- src/features/reducers.ts
@@ -1,7 +1,0 @@
-import { reducer as dictionary } from './dictionary/dictionary.slice';
-
-const reducers = {
-  dictionary,
-};
-
-export default reducers;
