+++ src/redux/dictionary/service.ts
@@ -1,20 +1,14 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
-import { getRawDefinitions } from './dictionary.api';
-
-interface IWordDetail {
-  bopomofo: string;
-  bopomofo2: string;
-  definitions: Array<string | Array<string>>;
-  pinyin: string;
-}
+import { getRawDefinitions } from './api';
+import { IWordDetail } from './dictionary';
 
 export const getDefinitions = createAsyncThunk('dictionary/getDefinitions', async (word: string) => {
   const { data } = await getRawDefinitions(word);
   return data.heteronyms.map((wordDetail: IWordDetail) => {
-    const definitionTypes = [...new Set(wordDetail.definitions.map((def: any) => def.type))];
+    const definitionTypes = [...new Set(wordDetail.definitions.map((def) => def.type))];
     const typeDefinitions = definitionTypes.map((type) => [
-      ...wordDetail.definitions.filter((def: any) => def.type === type),
+      ...wordDetail.definitions.filter((def) => def.type === type),
     ]);
     const pinyin = wordDetail.pinyin.split(' ');
