+++ src/redux/reducers.ts
@@ -1,0 +1,7 @@
+import { reducer as dictionary } from './dictionary/slice';
+
+const reducers = {
+  dictionary,
+};
+
+export default reducers;
